Change default C++ standard
The isAscii() and toAscii() functions aren't part of the C++14 standard,
though some distributions ship GCC binaries compiled such that it will
seem to be the case (i.e. std=c++14 will compile code where these
functions are called). The portable and "correct" solution is to use the
GNU14 standard instead. This is required for builds on for example
Alpine Linux.